Collapse the DetectorMiddleware into the SessionCookieMiddleware
